requirements.txt
-e .
coverage
mypy
pandas-stubs
pre-commit
pyright
pytest
textual[dev]
ipdb
pdbpp
